Move custom domain resources to dedicated pkg
